Split metrics.py into one file per class
