+++ src/modsim2/similarity/constants.py
@@ -1,8 +1,8 @@
-from modsim2.similarity import metrics
+from modsim2.similarity.metrics import mmd, otdd
 
 ARGUMENTS = "arguments"
 CLASS_KEY = "class"
 METRIC_CLS_DICT = {
-    "mmd": metrics.MMD,
-    "otdd": metrics.OTDD,
+    "mmd": mmd.MMD,
+    "otdd": otdd.OTDD,
 }
